version 007

+++ src/test/java/org/yorm/records/Person.java
@@ -2,7 +2,7 @@ package org.yorm.records;
 
 import java.time.LocalDateTime;
 
-public record Person(long id, String name, String email, LocalDateTime lastLogin, int companyId) {
+public record Person(int id, String name, String email, LocalDateTime lastLogin, int companyId) {
 
     public Person(String name) {
         this(0, name, null, null, 0);
